+++ src/components/Button/StoryButtonWrapper.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { ButtonSize } from "../global-types";
 import Button from "./Button";
 import { ButtonIconPosition, VariantType } from "./types";
 
@@ -7,14 +8,17 @@ export type StoryButtonWrapperProps = {
     disabled?: boolean;
     busy?: boolean;
     display?: boolean;
-    icon?: "trash" | "question" | "check" | "pen" | "car" | "car-side" | "cart-shopping" | "cart-plus";
+    icon?: "none" | "trash" | "question" | "check" | "pen" | "car" | "car-side" | "cart-shopping" | "cart-plus";
     iconPosition?: `${ButtonIconPosition}`;
     variant?: VariantType;
     tooltip?: string;
     disabledTooltip?: string;
+    size?: ButtonSize;
 };
 
 // This component is created only for storybook display purpose, i wanted to hide some of the props.
-const StoryButtonWrapper = (props: StoryButtonWrapperProps) => <Button {...props} onClick={() => {}} text="Button" />;
+const StoryButtonWrapper = ({ icon, ...other }: StoryButtonWrapperProps) => (
+    <Button {...other} icon={icon === "none" ? undefined : icon} onClick={() => {}} />
+);
 
 export default StoryButtonWrapper;