+++ src/components/Inputs/Slider/Slider.stories.tsx
@@ -5,7 +5,7 @@ import ThemeWrapper from "../../ThemeWrapper/ThemeWrapper";
 import Slider from "./Slider";
 
 export default {
-    title: "Morti-component-library/Inputs",
+    title: "M-component-library/Inputs",
     component: Slider,
 } as ComponentMeta<typeof Slider>;