+++ src/helpers/index.ts
@@ -1,3 +1,4 @@
 export * from "./getCssProperty";
+export * from "./getCssPropertyAsNumber";
 export * from "./getPosition";
 export * from "./setCssProperty";